Fire callback & add to history only after confirmation height is set
