+++ nano/core_test/CMakeLists.txt
@@ -9,6 +9,7 @@ add_executable (core_test
 	active_transactions.cpp
 	block.cpp
 	block_store.cpp
+	confirmation_height.cpp
 	conflicts.cpp
 	difficulty.cpp
 	entry.cpp
