+++ nano/node/active_transactions.cpp
@@ -103,6 +103,52 @@ void nano::active_transactions::confirm_frontiers (nano::transaction const & tra
 		next_frontier_check = steady_clock::now () + (representative_factor * fully_confirmed_factor / test_network_factor);
 	}
 }
+void nano::active_transactions::post_confirmation_height_set (nano::transaction const & transaction_a, std::shared_ptr<nano::block> block_a, nano::block_sideband const & sideband_a, nano::election_status_type election_status_type_a)
+{
+	if (election_status_type_a == nano::election_status_type::inactive_confirmation_height)
+	{
+		nano::account account (0);
+		nano::uint128_t amount (0);
+		bool is_state_send (false);
+		nano::account pending_account (0);
+		node.process_confirmed_data (transaction_a, block_a, block_a->hash (), sideband_a, account, amount, is_state_send, pending_account);
+		node.observers.blocks.notify (nano::election_status{ block_a, 0, std::chrono::duration_cast<std::chrono::milliseconds> (std::chrono::system_clock::now ().time_since_epoch ()), std::chrono::duration_values<std::chrono::milliseconds>::zero (), nano::election_status_type::inactive_confirmation_height }, account, amount, is_state_send);
+	}
+	else
+	{
+		auto hash (block_a->hash ());
+		std::unique_lock<std::mutex> lock (mutex);
+		auto existing (blocks.find (hash));
+		if (existing != blocks.end ())
+		{
+			auto election = existing->second;
+			if (election->confirmed && !election->stopped && election->status.winner->hash () == hash)
+			{
+				add_confirmed (existing->second->status, block_a->qualified_root ());
+
+				node.receive_confirmed (transaction_a, block_a, hash);
+				nano::account account (0);
+				nano::uint128_t amount (0);
+				bool is_state_send (false);
+				nano::account pending_account (0);
+				node.process_confirmed_data (transaction_a, block_a, hash, sideband_a, account, amount, is_state_send, pending_account);
+				election->status.type = election_status_type_a;
+				node.observers.blocks.notify (election->status, account, amount, is_state_send);
+				if (amount > 0)
+				{
+					node.observers.account_balance.notify (account, false);
+					if (!pending_account.is_zero ())
+					{
+						node.observers.account_balance.notify (pending_account, true);
+					}
+				}
+
+				election->clear_blocks ();
+				election->clear_dependent ();
+			}
+		}
+	}
+}
 
 void nano::active_transactions::request_confirm (std::unique_lock<std::mutex> & lock_a)
 {
@@ -129,11 +175,10 @@ void nano::active_transactions::request_confirm (std::unique_lock<std::mutex> & 
 		auto election_l (i->election);
 		if ((election_l->confirmed || election_l->stopped) && election_l->confirmation_request_count >= minimum_confirmation_request_count - 1)
 		{
-			if (election_l->confirmed)
+			if (election_l->stopped)
 			{
-				add_confirmed (election_l->status, root);
+				inactive.insert (root);
 			}
-			inactive.insert (root);
 		}
 		else
 		{
@@ -836,27 +881,27 @@ bool nano::active_transactions::publish (std::shared_ptr<nano::block> block_a)
 	return result;
 }
 
-void nano::active_transactions::confirm_block (nano::transaction const & transaction_a, std::shared_ptr<nano::block> block_a, nano::block_sideband const & sideband_a)
+// Returns the type of election status requiring callbacks calling later
+boost::optional<nano::election_status_type> nano::active_transactions::confirm_block (nano::transaction const & transaction_a, std::shared_ptr<nano::block> block_a, nano::block_sideband const & sideband_a)
 {
 	auto hash (block_a->hash ());
-	std::unique_lock<std::mutex> lock (mutex);
+	std::lock_guard<std::mutex> guard (mutex);
 	auto existing (blocks.find (hash));
 	if (existing != blocks.end ())
 	{
 		if (!existing->second->confirmed && !existing->second->stopped && existing->second->status.winner->hash () == hash)
 		{
 			existing->second->confirm_once (nano::election_status_type::active_confirmation_height);
+			return nano::election_status_type::active_confirmation_height;
+		}
+		else
+		{
+			return boost::optional<nano::election_status_type>{};
 		}
 	}
 	else
 	{
-		lock.unlock ();
-		nano::account account (0);
-		nano::uint128_t amount (0);
-		bool is_state_send (false);
-		nano::account pending_account (0);
-		node.process_confirmed_data (transaction_a, block_a, hash, sideband_a, account, amount, is_state_send, pending_account);
-		node.observers.blocks.notify (nano::election_status{ block_a, 0, std::chrono::duration_cast<std::chrono::milliseconds> (std::chrono::system_clock::now ().time_since_epoch ()), std::chrono::duration_values<std::chrono::milliseconds>::zero (), nano::election_status_type::inactive_confirmation_height }, account, amount, is_state_send);
+		return nano::election_status_type::inactive_confirmation_height;
 	}
 }
