+++ nano/node/confirmation_height_processor.cpp
@@ -4,6 +4,7 @@
 #include <nano/lib/utility.hpp>
 #include <nano/node/active_transactions.hpp>
 #include <nano/node/confirmation_height_processor.hpp>
+#include <nano/node/election.hpp>
 #include <nano/node/write_database_queue.hpp>
 #include <nano/secure/blockstore.hpp>
 #include <nano/secure/common.hpp>
@@ -49,7 +50,7 @@ void nano::confirmation_height_processor::run ()
 	std::unique_lock<std::mutex> lk (pending_confirmations.mutex);
 	while (!stopped)
 	{
-		if (!pending_confirmations.pending.empty ())
+		if (!paused && !pending_confirmations.pending.empty ())
 		{
 			pending_confirmations.current_hash = *pending_confirmations.pending.begin ();
 			pending_confirmations.pending.erase (pending_confirmations.current_hash);
@@ -82,6 +83,17 @@ void nano::confirmation_height_processor::run ()
 			}
 		}
 	}
+}
+
+void nano::confirmation_height_processor::pause ()
+{
+	paused = true;
+}
+
+void nano::confirmation_height_processor::unpause ()
+{
+	paused = false;
+	condition.notify_one ();
 }
 
 void nano::confirmation_height_processor::add (nano::block_hash const & hash_a)
@@ -147,6 +159,7 @@ void nano::confirmation_height_processor::add_confirmation_height (nano::block_h
 		}
 
 		auto count_before_receive = receive_source_pairs.size ();
+		std::vector<callback_data> block_callbacks_required;
 		if (block_height > iterated_height)
 		{
 			if ((block_height - iterated_height) > 20000)
@@ -154,7 +167,7 @@ void nano::confirmation_height_processor::add_confirmation_height (nano::block_h
 				logger.always_log ("Iterating over a large account chain for setting confirmation height. The top block: ", current.to_string ());
 			}
 
-			collect_unconfirmed_receive_and_sources_for_account (block_height, iterated_height, current, account, read_transaction);
+			collect_unconfirmed_receive_and_sources_for_account (block_height, iterated_height, current, account, read_transaction, block_callbacks_required);
 		}
 
 		// Exit early when the processor has been stopped, otherwise this function may take a
@@ -188,7 +201,7 @@ void nano::confirmation_height_processor::add_confirmation_height (nano::block_h
 					confirmed_iterated_pairs.emplace (account, confirmed_iterated_pair{ block_height, block_height });
 				}
 
-				pending_writes.emplace_back (account, current, block_height, block_height - confirmation_height);
+				pending_writes.emplace_back (account, current, block_height, block_height - confirmation_height, block_callbacks_required);
 			}
 
 			if (receive_details)
@@ -299,6 +312,11 @@ bool nano::confirmation_height_processor::write_pending (std::deque<conf_height_
 					return true;
 				}
 
+				for (auto & callback_data : pending.block_callbacks_required)
+				{
+					active.post_confirmation_height_set (transaction, callback_data.block, callback_data.sideband, callback_data.election_status_type);
+				}
+
 				stats.add (nano::stat::type::confirmation_height, nano::stat::detail::blocks_confirmed, nano::stat::dir::in, pending.height - confirmation_height);
 				assert (pending.num_blocks_confirmed == pending.height - confirmation_height);
 				confirmation_height = pending.height;
@@ -319,7 +337,7 @@ bool nano::confirmation_height_processor::write_pending (std::deque<conf_height_
 	return false;
 }
 
-void nano::confirmation_height_processor::collect_unconfirmed_receive_and_sources_for_account (uint64_t block_height_a, uint64_t confirmation_height_a, nano::block_hash const & hash_a, nano::account const & account_a, nano::read_transaction const & transaction_a)
+void nano::confirmation_height_processor::collect_unconfirmed_receive_and_sources_for_account (uint64_t block_height_a, uint64_t confirmation_height_a, nano::block_hash const & hash_a, nano::account const & account_a, nano::read_transaction const & transaction_a, std::vector<callback_data> & block_callbacks_required)
 {
 	auto hash (hash_a);
 	auto num_to_confirm = block_height_a - confirmation_height_a;
@@ -335,8 +353,18 @@ void nano::confirmation_height_processor::collect_unconfirmed_receive_and_source
 		{
 			if (!pending_confirmations.is_processing_block (hash))
 			{
-				active.confirm_block (transaction_a, block, sideband);
+				auto election_status_type = active.confirm_block (transaction_a, block, sideband);
+				if (election_status_type.is_initialized ())
+				{
+					block_callbacks_required.emplace_back (block, sideband, *election_status_type);
+				}
 			}
+			else
+			{
+				// This block is the original which is having its confirmation height set on
+				block_callbacks_required.emplace_back (block, sideband, nano::election_status_type::active_confirmed_quorum);
+			}
+
 			auto source (block->source ());
 			if (source.is_zero ())
 			{
@@ -350,9 +378,15 @@ void nano::confirmation_height_processor::collect_unconfirmed_receive_and_source
 				if (next_height != height_not_set)
 				{
 					receive_source_pairs.back ().receive_details.num_blocks_confirmed = next_height - block_height;
+
+					auto & receive_callbacks_required = receive_source_pairs.back ().receive_details.block_callbacks_required;
+
+					// Don't include the last one as that belongs to the next recieve
+					std::copy (block_callbacks_required.begin (), block_callbacks_required.end () - 1, std::back_inserter (receive_callbacks_required));
+					block_callbacks_required = { block_callbacks_required.back () };
 				}
 
-				receive_source_pairs.emplace_back (conf_height_details{ account_a, hash, block_height, height_not_set }, source);
+				receive_source_pairs.emplace_back (conf_height_details{ account_a, hash, block_height, height_not_set, {} }, source);
 				++receive_source_pairs_size;
 				next_height = block_height;
 			}
@@ -374,16 +408,18 @@ void nano::confirmation_height_processor::collect_unconfirmed_receive_and_source
 	{
 		auto & last_receive_details = receive_source_pairs.back ().receive_details;
 		last_receive_details.num_blocks_confirmed = last_receive_details.height - confirmation_height_a;
+		last_receive_details.block_callbacks_required = block_callbacks_required;
 	}
 }
 
 namespace nano
 {
-confirmation_height_processor::conf_height_details::conf_height_details (nano::account const & account_a, nano::block_hash const & hash_a, uint64_t height_a, uint64_t num_blocks_confirmed_a) :
+confirmation_height_processor::conf_height_details::conf_height_details (nano::account const & account_a, nano::block_hash const & hash_a, uint64_t height_a, uint64_t num_blocks_confirmed_a, std::vector<callback_data> const & block_callbacks_required_a) :
 account (account_a),
 hash (hash_a),
 height (height_a),
-num_blocks_confirmed (num_blocks_confirmed_a)
+num_blocks_confirmed (num_blocks_confirmed_a),
+block_callbacks_required (block_callbacks_required_a)
 {
 }
 
@@ -395,6 +431,13 @@ source_hash (source_a)
 
 confirmation_height_processor::confirmed_iterated_pair::confirmed_iterated_pair (uint64_t confirmed_height_a, uint64_t iterated_height_a) :
 confirmed_height (confirmed_height_a), iterated_height (iterated_height_a)
+{
+}
+
+confirmation_height_processor::callback_data::callback_data (std::shared_ptr<nano::block> const & block_a, nano::block_sideband const & sideband_a, nano::election_status_type election_status_type_a) :
+block (block_a),
+sideband (sideband_a),
+election_status_type (election_status_type_a)
 {
 }
