+++ nano/node/confirmation_height_processor.hpp
@@ -1,6 +1,8 @@
 #pragma once
 
 #include <nano/lib/numbers.hpp>
+#include <nano/node/active_transactions.hpp>
+#include <nano/secure/blockstore.hpp>
 #include <nano/secure/common.hpp>
 
 #include <condition_variable>
@@ -31,6 +33,7 @@ private:
 	nano::block_hash current_hash{ 0 };
 	friend class confirmation_height_processor;
 	friend class confirmation_height_pending_observer_callbacks_Test;
+	friend class confirmation_height_dependent_election_Test;
 };
 
 std::unique_ptr<seq_con_info_component> collect_seq_con_info (pending_confirmation_height &, const std::string &);
@@ -42,6 +45,8 @@ public:
 	~confirmation_height_processor ();
 	void add (nano::block_hash const &);
 	void stop ();
+	void pause ();
+	void unpause ();
 
 	/** The maximum amount of accounts to iterate over while writing */
 	static uint64_t constexpr batch_write_size = 2048;
@@ -50,15 +55,25 @@ public:
 	static uint64_t constexpr batch_read_size = 4096;
 
 private:
+	class callback_data final
+	{
+	public:
+		callback_data (std::shared_ptr<nano::block> const &, nano::block_sideband const &, nano::election_status_type);
+		std::shared_ptr<nano::block> block;
+		nano::block_sideband sideband;
+		nano::election_status_type election_status_type;
+	};
+
 	class conf_height_details final
 	{
 	public:
-		conf_height_details (nano::account const &, nano::block_hash const &, uint64_t, uint64_t);
+		conf_height_details (nano::account const &, nano::block_hash const &, uint64_t, uint64_t, std::vector<callback_data> const &);
 
 		nano::account account;
 		nano::block_hash hash;
 		uint64_t height;
 		uint64_t num_blocks_confirmed;
+		std::vector<callback_data> block_callbacks_required;
 	};
 
 	class receive_source_pair final
@@ -81,6 +96,7 @@ private:
 	std::condition_variable condition;
 	nano::pending_confirmation_height & pending_confirmations;
 	std::atomic<bool> stopped{ false };
+	std::atomic<bool> paused{ false };
 	nano::block_store & store;
 	nano::stat & stats;
 	nano::active_transactions & active;
@@ -100,7 +116,7 @@ private:
 
 	void run ();
 	void add_confirmation_height (nano::block_hash const &);
-	void collect_unconfirmed_receive_and_sources_for_account (uint64_t, uint64_t, nano::block_hash const &, nano::account const &, nano::read_transaction const &);
+	void collect_unconfirmed_receive_and_sources_for_account (uint64_t, uint64_t, nano::block_hash const &, nano::account const &, nano::read_transaction const &, std::vector<callback_data> &);
 	bool write_pending (std::deque<conf_height_details> &);
 
 	friend std::unique_ptr<seq_con_info_component> collect_seq_con_info (confirmation_height_processor &, const std::string &);
