+++ nano/node/election.cpp
@@ -50,11 +50,7 @@ void nano::election::confirm_once (nano::election_status_type type_a)
 		{
 			--node.active.long_unconfirmed_size;
 		}
-		auto root (status.winner->qualified_root ());
-		node.active.add_confirmed (status, root);
-		clear_blocks ();
-		clear_dependent ();
-		node.active.roots.erase (root);
+		node.active.roots.erase (status.winner->qualified_root ());
 	}
 }
